lambda/requirements.txt
ask-sdk==1.10.2
ask-sdk-core>=1.10.2
ask-sdk-dynamodb-persistence-adapter==1.10.2
ask-sdk-model==1.29.0
Pyrogram==1.0.7
pytz==2020.1
